Missing Translation of Scooter
Fix some miss translation. The icon scooter as it is in the traccar, that was uploaded by me is to a Eletric Scooter used by Grin, Yellow, Bird, Lime. It is a different scooter than a "Lambreta", "Lambreta" is more like a motocycle with less motor power. It's also called Scooter in english, but it's nothing looking like the icon I uploaded. Not the